==> Rscripts/6.0.CreateFastaAlignment.R <==
#Create a fasta file for each monkey with PID concensus sequences

library(reshape)
library(msa)
library(bios2mds)
library(DataCombine)
library(GenomicAlignments)
source("Rscripts/baseRscript.R")

bams<-list.files("Output/bam_PID/",pattern="^Run8.+bam$")

#parameter setting for stackStringsFromBam
gr <- GRanges(seqnames="AY032751env",  IRanges(215, 691))

#Create an alignment file from bam files pos 215:691
for (i in 1:length(bams)){
    index<-paste0("Output/bam_PID/",bams[i],'.bai')
    bamfile <- BamFile(paste0("Output/bam_PID/",bams[i]), index=index)

    fname<-substr(bams[i], 1,7)
    seq<-stackStringsFromBam(bamfile, param=gr)
    
    #write the alignment fasta file
    writeXStringSet(seq,paste0("Output/PID_Con_Alignment/", fname, ".Alignment.fasta"),width =500)
}
    
### Replace '+' with '-' for the fasta files saved above in terminal
### cd to "Output" directory
### >     LC_ALL=C find PID_Con_Alignment -type f -exec sed -i "" 's/+/-/g' {} \;

#For Run2 files, use a different method to create an alignment files.
# Currently the sequences with end portion only were simply removed
